requirements.txt
Flask==3.0.3
openai==0.27.8
python-dotenv==1.0.1
Requests==2.32.2
